Modify RoomsNavigationComponent to work with the new hooks
